Indexer refactoring: merged MultithreadedIndexer into IndexCollection.
